// client/src/app/models/Entity/products.model.ts
export interface IProducts{
    id?: Number;
    name?: String;
    price?: Number;
    sellPrice?: Number;
    inputPrice?: Number;
    description?: String;
    purchases?: Number;
    idCatalog?: Number;
    inventory?: Number;
    idCompany?: Number;
    userCreate?: Number;
    quantitys?: Number;
}